Approximate Square Root Calculation Using the Babylonian Method
**Procedure:**  
1. **Input: x>=0
2. **Output:** Approximation of the square root of \( x \)  

3. \( t = x \)  
4. **while** (t is still far from the square root of x):  
5. → \( t = \frac{1}{2} \cdot (t + \frac{x}{t}) \)  
6. **return** t  

In the task above, you implemented functions that take exactly one parameter and return a double value. The calculation of the return value involves a (predefined) tolerance value used in the loop's termination condition.

Now, overload each of the previously defined functions so that the tolerance value can be specified by the user.